pyproject.toml
[project]
name = "earlybird"
version = "0.1.0"
description = "AI/tech content radar — scrape, filter, deliver"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "pydantic>=2.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "umap-learn>=0.5",
    "igraph>=0.11",
    "leidenalg>=0.10",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "shapely>=2.0",
]

[tool.setuptools.packages.find]
include = ["earlybird*"]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio"]

[project.scripts]
earlybird-scrape = "earlybird.scraper:main"
earlybird-feed = "earlybird.build_daily_feed:main"
